// src/store/modules/transfers/enums.ts
export enum MutationsNames {
  SET_TRANSFERS = "SET_TRANSFERS",
  CHANGE_FILTER = "CHANGE_FILTER",
  SET_CATEGORY = "SET_CATEGORY"
}

export enum ActionsNames {
  GET_TRANSFERS = "GET_TRANSFERS",
  ADD_TRANSFER = "ADD_TRANSFER",
  SET_CATEGORY = "SET_CATEGORY",
  DELETE_TRANSFER = "DELETE_TRANSFER"
}